+++ DataTransferObject/Requests/DTOProfileRequest.cs
@@ -31,7 +31,7 @@ namespace DataTransferObject.Requests
         public byte FormationId { get; set; }
 
         [Required(ErrorMessageResourceType = typeof(ErrorMessages), ErrorMessageResourceName = "RequiredError")]
-        public byte ApptId { get; set; }
+        public short ApptId { get; set; }
         public MAppointment? MAppointment { get; set; }
         public Boolean IntOffr { get; set; }
     }
